# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/integrations/omie/connection": {
      "get": {
        "operationId": "OmieController_getConnection",
        "summary": "Exibir status da conexão Omie (ADMIN)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "integrations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "OmieController_upsertConnection",
        "summary": "Atualizar credenciais Omie (ADMIN)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertOmieConnectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "integrations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/integrations/omie/connection/test": {
      "post": {
        "operationId": "OmieController_testConnection",
        "summary": "Testar credenciais Omie (ADMIN)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestOmieConnectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "integrations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/integrations/omie/events": {
      "get": {
        "operationId": "OmieController_listEvents",
        "summary": "Listar eventos Omie por tenant (ADMIN)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "integrations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/integrations/omie/reprocess/{eventId}": {
      "post": {
        "operationId": "OmieController_reprocessEvent",
        "summary": "Reprocessar evento Omie com erro (ADMIN)",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "integrations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/integrations/omie/internal/process/{eventId}": {
      "post": {
        "operationId": "OmieController_processEvent",
        "summary": "Processar evento Omie (chamado pelo worker)",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "integrations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Efizion Bath API",
    "description": "Contratos REST padronizados com envelope `{ data, meta }` para todas as respostas. Use os exemplos fornecidos para alinhar payloads entre frontend e backend.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.efizion.com.br",
      "description": "Produção"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Tokens emitidos pelos fluxos /v1/auth ou /v1/customer-auth"
      },
      "x-request-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-request-id",
        "description": "Opcional. Permite correlação distribuída por request."
      }
    },
    "schemas": {
...
      "UpsertOmieConnectionDto": {
        "type": "object",
        "properties": {}
      },
      "TestOmieConnectionDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}